+++ docs/release-notes.md
@@ -36,6 +36,7 @@ hide:
 - Invalidating caused schema errors.
 - ManyToMany and ForeignKey fields didn't worked when referencing tenant models.
 - ManyToMany fields didn't worked when specified on tenant models.
+- Fix transaction method to work on instance and class.
 
 ### BREAKING